Backend/models/src/models/CourseNotes.java
package models;

public class CourseNotes extends CrossReferencingObject {

    private String title;
    private String url;
    private int id;

    public CourseNotes(String title, String url) {
        super();
        this.title = title;
        this.url = url;
    }

    public String getTitle() {
        return title;
    }

    public void setTitle(String title) {
        this.title = title;
    }

    public String getUrl() {
        return url;
    }

    public void setUrl(String url) {
        this.url = url;
    }

    public int getId() {
        return id;
    }

    public void setId(int id) {
        this.id = id;
    }

}
